+++ src/commands/vox.rs
@@ -1,12 +1,10 @@
-use crate::{
-    dectalk::{Dectalk, WaveFormat},
-    types::{Context, Error},
-};
+use crate::dectalk::{Dectalk, WaveFormat};
+use crate::pawthos::types::{Context, Result};
 use poise::serenity_prelude as serenity;
 
 /// Vox: A suite of commands around Dectalk voice synthesis.
 #[poise::command(slash_command, subcommands("say"), subcommand_required)]
-pub async fn vox(_: Context<'_>) -> Result<(), Error> {
+pub async fn vox(_: Context<'_>) -> Result {
     //lmao, again
     panic!();
 }
@@ -17,8 +15,8 @@ pub async fn say(
     ctx: Context<'_>,
     #[description = "Text to synthesize"] text: String,
     #[description = "Words per minute"] rate: Option<u32>,
-) -> Result<(), Error> {
-    let path = tokio::task::spawn_blocking(move || -> Result<std::path::PathBuf, Error> {
+) -> Result {
+    let path = tokio::task::spawn_blocking(move || -> Result<std::path::PathBuf> {
         use std::{
             env,
             time::{SystemTime, UNIX_EPOCH},
