+++ src/pawthos/enums/mod.rs
@@ -1,0 +1,3 @@
+pub mod embed_type;
+pub mod logos_errors;
+pub mod persistant_data;
