+++ src/pawthos/enums/persistant_data.rs
@@ -1,0 +1,6 @@
+use crate::pawthos::structs::mimic_db::MimicDB;
+
+#[derive(Debug)]
+pub enum PersistantData {
+    MimicDB(MimicDB),
+}
