+++ src/pawthos/structs/mod.rs
@@ -1,0 +1,4 @@
+pub mod data;
+pub mod mimic;
+pub mod mimic_db;
+pub mod mimic_user;
